explore minimal dropzone
