+++ packages/dropzone/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "@uploadthing/dropzone",
+  "version": "0.1.0",
+  "license": "MIT",
+  "exports": {
+    "./package.json": "./package.json",
+    "./core": {
+      "types": "./dist/core.d.ts",
+      "import": "./dist/core.mjs",
+      "require": "./dist/core.js"
+    },
+    "./react": {
+      "types": "./dist/react.d.ts",
+      "import": "./dist/react.mjs",
+      "require": "./dist/react.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "dist/*"
+      ]
+    }
+  },
+  "scripts": {
+    "lint": "eslint \"**/*.{ts,tsx}\" --max-warnings 0",
+    "build": "tsup",
+    "clean": "git clean -xdf dist node_modules",
+    "dev": "tsup --watch",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "attr-accept": "^2.2.2",
+    "file-selector": "^0.6.0"
+  },
+  "devDependencies": {
+    "@types/react": "18.2.33",
+    "@uploadthing/eslint-config": "0.2.0",
+    "@uploadthing/tsconfig": "0.1.0",
+    "@uploadthing/tsup-config": "0.1.0",
+    "eslint": "^8.47.0",
+    "react": "18.2.0",
+    "tsup": "6.7.0",
+    "typescript": "^5.1.6"
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "@uploadthing/eslint-config/base"
+    ]
+  }
+}
