+++ packages/dropzone/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@uploadthing/tsconfig/base.json",
+  "include": ["**/*.ts", "**/*.tsx"]
+}
