+++ packages/react/package.json
@@ -44,6 +44,7 @@
     "dev": "wait-on ../uploadthing/dist/server.mjs && tsup --watch"
   },
   "dependencies": {
+    "@uploadthing/dropzone": "^0.1.0",
     "@uploadthing/shared": "^6.0.2",
     "attr-accept": "^2.2.2",
     "file-selector": "^0.6.0",
