+++ packages/react/src/hooks.ts
@@ -1,3 +1,3 @@
 export { generateReactHelpers } from "./useUploadThing";
 
-export { useDropzone } from "./use-dropzone";
+export { useDropzone } from "@uploadthing/dropzone/react";
